Refactor adapt grid state in cache service
